pyproject.toml
[tool.poetry]
name = "nsp2-ultrasound"
version = "0.1.0"
description = ""
authors = ["MissaFissa <meho11@outlook.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
nump = "^5.5.5.5"
scipy = "^1.14.1"
matplotlib = "^3.9.2"
click = "^8.1.7"
pandas = "^2.2.3"
plotly = "^5.24.1"


[tool.poetry.group.dev.dependencies]
poethepoet = "^0.30.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
run = "nsp2_ultrasound.cli:initiate"
plot = "nsp2_ultrasound.gui:main"